requirements.txt
# API e Web
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.4.2
email-validator==2.1.0.post1

# Processamento de Dados
pandas==2.1.3
numpy==1.26.2

# Processamento de Imagem
opencv-python==4.8.1.78
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.196    # YOLO

# Cache e Performance
cachetools==5.3.2
prometheus-client==0.19.0
psutil==5.9.6

# Logging e Monitoramento
python-json-logger==2.0.7
structlog==23.2.0
prometheus-client==0.19.0

# Testes
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-mock==3.12.0

# Utilitários
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2023.3.post1

# Segurança
cryptography==41.0.5
pyjwt==2.8.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Documentação
mkdocs==1.5.3
mkdocs-material==9.4.14
mkdocstrings==0.23.0
mkdocstrings-python==1.7.1

# Desenvolvimento
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0 